refactor: split world npc persistence owners
